+++ functions/private.ps1
@@ -1,11 +1,16 @@
 #these are private helper functions
 
 function pause {
-    [void](Read-Host "`e[3;37m$($strings.continue)$($PSStyle.Reset)")
-    "`e[2A"
-    "`e[K"
-    "`e[3A"
-    "`e[1G"
+    $r=(Read-Host "`e[3;37m$($strings.continue)$($PSStyle.Reset)")
+    if ($r -eq 'q') {
+        Break
+    }
+    else {
+        "`e[2A"
+        "`e[K"
+        "`e[3A"
+        "`e[1G"
+    }
 }
 
 Function Format-BorderBox {