+++ CHANGELOG.md
@@ -4,6 +4,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Added online help links for English help files.
+- Added a welcome style to the PSIntro screen
+
+### Changed
+- Revised `Get-ModuleStatus` to use `Find-PSResource` and not `Find-Module`.
 
 ## [0.9.0] - 2025-06-17